David/20240227 rate routes a (#67)
Route rating in gym routes view - 16h

This PR contains new feature where it's possible to rate routes directly in the gym routes view. This is enabled only for logged in users. A new grade '?' was added and it can be used when first opening a route and then users can rate the route however they feel. It can be used also for already graded routes to perhaps suggest another grade.

A rating is added to user's daily activity under name Rating activity. If there is an activity already, the route score will be added to the existing today's activity. For now it's all mixed in but in the end it is the same thing whether a user rates a route differently when adding it in the Activities page or in the gym routes view.